+++ internal/command/init.go
@@ -31,6 +31,7 @@ import (
 	"github.com/opentofu/opentofu/internal/states"
 	"github.com/opentofu/opentofu/internal/tfdiags"
 	"github.com/opentofu/opentofu/internal/tofu"
+	"github.com/opentofu/opentofu/internal/tofumigrate"
 	tfversion "github.com/opentofu/opentofu/version"
 )
 
@@ -300,6 +301,18 @@ func (c *InitCommand) Run(args []string) int {
 				return 1
 			}
 		}
+	}
+
+	if state != nil {
+		// Since we now have the full configuration loaded, we can use it to migrate the in-memory state view
+		// prior to fetching providers.
+		migratedState, migrateDiags := tofumigrate.MigrateStateProviderAddresses(config, state)
+		diags = diags.Append(migrateDiags)
+		if migrateDiags.HasErrors() {
+			c.showDiagnostics(diags)
+			return 1
+		}
+		state = migratedState
 	}
 
 	// Now that we have loaded all modules, check the module tree for missing providers.